Move date controls into its own module
